+++ packages/transaction-pay-controller/src/strategy/across/across-quotes.ts
@@ -136,6 +136,7 @@ async function getSingleQuote(
   const originalQuote: AcrossQuoteWithoutMetaMask = {
     quote,
     request: {
+      actions: destination.actions,
       amount,
       tradeType,
     },